putting firebase auth in separate file
